fix: 🐛 PARAMETER_TYPE definition fix
