+++ packages/ogdf.js/src/utils/parameter-type.js
@@ -1,0 +1,8 @@
+
+export const PARAMETER_TYPE = {
+    MODULE: "MODULE",
+    CATEGORICAL: "CATEGORICAL",
+    INT: "INT",
+    DOUBLE: "DOUBLE",
+    BOOL: "BOOL",
+}
